+++ mqtt/NativeFormat.js
@@ -1,4 +1,4 @@
-class NativePayload {
+class NativeFormat {
     #log;
     #server;
 
@@ -9,22 +9,34 @@ class NativePayload {
         this.#server = platform.server;
     }
 
-    async getStatus() {
-        this.#log.debug(this.constructor.name, 'getStatus');
+    get subscriptionTopics() {
+        return ['native/set'];
+    }
+
+    async getStatusMessages() {
+        this.#log.debug(this.constructor.name, 'getStatusMessages');
 
         try {
             const status = await this.#server.getStatus();
-            return status.toString();
+            return [{
+                topic: 'native/get',
+                payload: status.toString()
+            }];
         }
         catch(error) {
             this.#log.error(error);
         }
     }
 
-    async processCommand(payload) {
+    async processCommandMessage(topic, payload) {
         this.#log.debug(this.constructor.name, 'processCommand', payload);
 
         try {
+            if (topic !== 'native/set') {
+                this.#log.warn(`Invalid topic in MQTT message: ${topic}`);
+                return;
+            }
+
             let json = JSON.parse(payload);
 
             let group1 = Object.keys(json)[0];
@@ -51,4 +63,4 @@ class NativePayload {
     }
 }
 
-module.exports = NativePayload;
+module.exports = NativeFormat;